requirements.txt
requests
psycopg2-binary==2.9.5